fix: address review on scoped enum migration
- SQLiteClientLogger_RNSLogger.cpp: add (int) casts to scoped-enum array
  subscripts (this optional logger TU isn't built in the default config, so the
  rename introduced a latent compile error the build didn't catch).
- peer.h / peerinterface.h: drop the stale 'set priority to 0 to suppress
  ID_DISCONNECTION_NOTIFICATION' note — never true (0 == Immediate, the highest
  priority, and the notification is sent unconditionally), and impossible now
  that the parameter is a scoped enum.
- MessageIdentifiers.h: reword the orphaned '_WITH_ACK_RECEIPT' doc reference to
  name the MafiaNet::Reliability ...WithAckReceipt values.

+++ DependentExtensions/SQLite3Plugin/Logger/ClientOnly/Optional/SQLiteClientLogger_RNSLogger.cpp
@@ -58,14 +58,14 @@ void SQLiteClientLogger_RakNetStatistics::Update(void)
 					rns.isLimitedByCongestionControl,
 					rns.BPSLimitByOutgoingBandwidthLimit,
 					rns.isLimitedByOutgoingBandwidthLimit,
-					rns.messageInSendBuffer[MafiaNet::Priority::Immediate],
-					rns.messageInSendBuffer[MafiaNet::Priority::High],
-					rns.messageInSendBuffer[MafiaNet::Priority::Medium],
-					rns.messageInSendBuffer[MafiaNet::Priority::Low],
-					rns.bytesInSendBuffer[MafiaNet::Priority::Immediate],
-					rns.bytesInSendBuffer[MafiaNet::Priority::High],
-					rns.bytesInSendBuffer[MafiaNet::Priority::Medium],
-					rns.bytesInSendBuffer[MafiaNet::Priority::Low],
+					rns.messageInSendBuffer[(int)MafiaNet::Priority::Immediate],
+					rns.messageInSendBuffer[(int)MafiaNet::Priority::High],
+					rns.messageInSendBuffer[(int)MafiaNet::Priority::Medium],
+					rns.messageInSendBuffer[(int)MafiaNet::Priority::Low],
+					rns.bytesInSendBuffer[(int)MafiaNet::Priority::Immediate],
+					rns.bytesInSendBuffer[(int)MafiaNet::Priority::High],
+					rns.bytesInSendBuffer[(int)MafiaNet::Priority::Medium],
+					rns.bytesInSendBuffer[(int)MafiaNet::Priority::Low],
 					rns.messagesInResendBuffer,
 					rns.bytesInResendBuffer,
 					rns.packetlossLastSecond,
@@ -100,7 +100,7 @@ void SQLiteClientLogger_RakNetStatistics::Update(void)
 					rns.valueOverLastSecond[ACTUAL_BYTES_RECEIVED], \
 					rns.BPSLimitByCongestionControl, \
 					rns.BPSLimitByOutgoingBandwidthLimit, \
-					rns.bytesInSendBuffer[MafiaNet::Priority::Immediate]+rns.bytesInSendBuffer[MafiaNet::Priority::High]+rns.bytesInSendBuffer[MafiaNet::Priority::Medium]+rns.bytesInSendBuffer[MafiaNet::Priority::Low], \
+					rns.bytesInSendBuffer[(int)MafiaNet::Priority::Immediate]+rns.bytesInSendBuffer[(int)MafiaNet::Priority::High]+rns.bytesInSendBuffer[(int)MafiaNet::Priority::Medium]+rns.bytesInSendBuffer[(int)MafiaNet::Priority::Low], \
 					rns.messagesInResendBuffer, \
 					rns.bytesInResendBuffer, \
 					rns.packetlossLastSecond, \
